Sync with upstream parser definition
Take the latest version of VisualBasic6.g4 from upstream project. Due to the extensive re-formatting of that file going to need to re-apply local changes so that a PR can be submitted.

For now, just get everything building and comment out those bits of code/tests which were reliant on previous local changes. These will be uncommented as the changes are re-applied.